requirements.txt
pyTelegramBotAPI==4.30.0
python-dotenv==1.2.1
